PI/interface/requirements.txt
luma.oled>=3.13
RPi.GPIO>=0.7
Pillow>=9.0
